edge/raspberry-pi/requirements.txt
# Lifo4 EMS - Raspberry Pi Edge Gateway

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# MQTT
paho-mqtt==2.0.0

# Modbus
pymodbus==3.6.0

# CAN bus
python-can==4.3.0

# Serial communication
pyserial==3.5

# Data processing
numpy==1.26.0
pandas==2.1.0

# Database (local storage)
sqlalchemy==2.0.0
aiosqlite==0.19.0

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Monitoring
prometheus-client==0.19.0

# GPIO (Raspberry Pi specific)
# RPi.GPIO==0.7.1  # Uncomment on Raspberry Pi
# gpiozero==2.0

# Scheduling
apscheduler==3.10.0

# HTTP client
httpx==0.26.0

# Logging
structlog==24.1.0
